Update Natura to 4.3.2.62; remove most recipe customizations
My hack to change recipe priority and fix the conflicts with vanilla
recipes was a little too disruptive -- it broke vanilla advancements,
for one, because I was removing the vanilla tool recipes from the game.
There is a new version of Natura that fixes the fence and fence gate
recipes by using the new vanilla recipe pattern, so accept that as good
enough and settle for removing Natura's varied sticks from the game.

There are still conflicts with vanilla recipes that can't be solved
because 1.12.2 Forge has no recipe priority system[1]. Add
NoMoreRecipeConflict to work around them.

1. https://github.com/MinecraftForge/MinecraftForge/issues/4415